third_party/blink/web_tests/http/tests/navigation/back-during-child-frame-reload-expected.txt

Navigation initiated by history.back() should overwrite child frame reload, which was started eralier and did not finish yet. History state at the end should reflect the back() call.

============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/back-during-child-frame-reload.html
            about:blank (in frame "<!--framePath //<!--frame0-->-->")
curr->  http://127.0.0.1:8000/navigation/back-during-child-frame-reload.html?one
            about:blank (in frame "<!--framePath //<!--frame0-->-->")
        http://127.0.0.1:8000/navigation/back-during-child-frame-reload.html?one?two
            about:blank (in frame "<!--framePath //<!--frame0-->-->")
===============================================
